+++ src/main/java/com/mageddo/dnsproxyserver/solver/SolverCacheFactory.java
@@ -72,9 +72,17 @@ public class SolverCacheFactory {
       ;
   }
 
+  /**
+   * This method should be called from one single thread, or it will cause deadlock.
+   */
   public void clearCaches() {
+    // fixme #526 possible solutions for the deadlock:
+    //       1 - only one thread can clear the cache at a time
+    //       2 - move the locks to one centralized thread responsible for the cache management
     for (final var cache : this.getCaches()) {
+      log.trace("status=clearing, cache={}", cache.name());
       cache.clear();
+      log.trace("status=cleared, cache={}", cache.name());
     }
   }
 }