+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/configurator/CircuitBreakerWatchDogScheduler.java
@@ -4,12 +4,14 @@ import com.mageddo.commons.concurrent.ThreadPool;
 import com.mageddo.dnsproxyserver.di.StartupEvent;
 import com.mageddo.dnsproxyserver.solver.remote.application.CircuitBreakerFactory;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
+@Slf4j
 @Singleton
 @RequiredArgsConstructor(onConstructor = @__({@Inject}))
 public class CircuitBreakerWatchDogScheduler implements StartupEvent {
@@ -19,6 +21,14 @@ public class CircuitBreakerWatchDogScheduler implements StartupEvent {
 
   @Override
   public void onStart() {
-    this.executor.scheduleWithFixedDelay(this.circuitBreakerFactory::checkCreatedCircuits, 0, 10, TimeUnit.SECONDS);
+    this.executor.scheduleWithFixedDelay(this::logStats, 0, 10, TimeUnit.SECONDS);
+  }
+
+  void logStats() {
+    this.circuitBreakerFactory
+      .stats()
+      .forEach(stats -> {
+        log.debug("stats={}", stats);
+      });
   }
 }