+++ src/test/java/com/mageddo/dnsproxyserver/solver/remote/application/CircuitBreakerFactoryTest.java
@@ -1,5 +1,8 @@
 package com.mageddo.dnsproxyserver.solver.remote.application;
 
+import com.mageddo.commons.circuitbreaker.CircuitCheckException;
+import com.mageddo.commons.concurrent.Threads;
+import com.mageddo.dnsproxyserver.solver.remote.Result;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -7,10 +10,17 @@ import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import testing.templates.CircuitBreakerConfigTemplates;
 import testing.templates.InetSocketAddressTemplates;
+import testing.templates.solver.remote.ResultSupplierTemplates;
+
+import java.net.InetSocketAddress;
+import java.util.function.Supplier;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 class CircuitBreakerFactoryTest {
@@ -29,7 +39,7 @@ class CircuitBreakerFactoryTest {
     doReturn(true).when(this.factory).circuitBreakerSafeCheck(any());
 
     final var addr = InetSocketAddressTemplates._8_8_8_8();
-    this.factory.createCircuitBreakerFor(addr);
+    this.factory.createOrGetCircuitBreaker(addr);
 
     // act
     final var result = this.factory.checkCreatedCircuits();
@@ -50,7 +60,7 @@ class CircuitBreakerFactoryTest {
     doReturn(false).when(this.factory).circuitBreakerSafeCheck(any());
 
     final var addr = InetSocketAddressTemplates._8_8_8_8();
-    this.factory.createCircuitBreakerFor(addr);
+    this.factory.createOrGetCircuitBreaker(addr);
 
     // act
     final var result = this.factory.checkCreatedCircuits();
@@ -58,5 +68,43 @@ class CircuitBreakerFactoryTest {
     // assert
     assertEquals(0, result.getKey());
     assertEquals(1, result.getValue());
+  }
+
+  @Test
+  void mustNotFlushCacheWhenChangeStateToHalfOpen(){
+
+    // arrange
+    assertEquals("[]", this.factory.stats().toString());
+
+    final var addr = InetSocketAddressTemplates._8_8_8_8();
+    final var supError = ResultSupplierTemplates.alwaysFail();
+    final var supSuccess = ResultSupplierTemplates.alwaysSuccess();
+
+    doReturn(CircuitBreakerConfigTemplates.oneTryFailSuccess())
+      .when(this.factory)
+      .findCircuitBreakerConfig()
+    ;
+
+    this.checkFailAndSleep(addr, supError);
+    this.checkFailAndSleep(addr, supError);
+
+    this.factory.check(addr, supSuccess);
+    assertEquals(
+      "[CircuitBreakerFactory.Stats(remoteServerAddress=/8.8.8.8:53, state=CLOSED)]",
+      this.factory.stats().toString()
+    );
+    verify(this.factory, times(2)).flushCache();
+
+
+  }
+
+  void checkFailAndSleep(InetSocketAddress addr, Supplier<Result> supError) {
+    assertThrows(CircuitCheckException.class, () -> this.factory.check(addr, supError));
+    assertEquals(
+      "[CircuitBreakerFactory.Stats(remoteServerAddress=/8.8.8.8:53, state=OPEN)]",
+      this.factory.stats().toString()
+    );
+    verify(this.factory).flushCache();
+    Threads.sleep(100);
   }
 }